refactor: move html elements to @/ui
